Add files via upload
This was written for a video game that I play / an exam. It's purpose is to take groups of three skills, look at the first skill, and then move forward with making sure that each of the three skills appears twice throughout x groupings while making sure that the first skill is not the same as other groups.